COMMERCE-11790 Use new search framework for CPDefinition
