+++ modules/apps/commerce/commerce-product-service/src/main/java/com/liferay/commerce/product/internal/search/spi/model/query/contributor/CPDefinitionModelPreFilterContributor.java
@@ -5,16 +5,24 @@
 
 package com.liferay.commerce.product.internal.search.spi.model.query.contributor;
 
+import com.liferay.commerce.product.constants.CPField;
 import com.liferay.commerce.product.model.CommerceCatalog;
 import com.liferay.commerce.product.service.CommerceCatalogService;
 import com.liferay.petra.function.transform.TransformUtil;
 import com.liferay.portal.kernel.dao.orm.QueryUtil;
 import com.liferay.portal.kernel.search.BooleanClauseOccur;
+import com.liferay.portal.kernel.search.Field;
 import com.liferay.portal.kernel.search.SearchContext;
+import com.liferay.portal.kernel.search.facet.util.RangeParserUtil;
 import com.liferay.portal.kernel.search.filter.BooleanFilter;
+import com.liferay.portal.kernel.search.filter.Filter;
+import com.liferay.portal.kernel.search.filter.RangeTermFilter;
+import com.liferay.portal.kernel.search.filter.TermFilter;
 import com.liferay.portal.kernel.search.filter.TermsFilter;
 import com.liferay.portal.kernel.util.ArrayUtil;
 import com.liferay.portal.kernel.util.GetterUtil;
+import com.liferay.portal.kernel.util.Validator;
+import com.liferay.portal.kernel.workflow.WorkflowConstants;
 import com.liferay.portal.search.spi.model.query.contributor.ModelPreFilterContributor;
 import com.liferay.portal.search.spi.model.registrar.ModelSearchSettings;
 
@@ -22,6 +30,7 @@ import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
 
 /**
+ * @author Brian I. Kim
  * @author Petteri Karttunen
  */
 @Component(
@@ -39,26 +48,236 @@ public class CPDefinitionModelPreFilterContributor
 		BooleanFilter booleanFilter, ModelSearchSettings modelSearchSettings,
 		SearchContext searchContext) {
 
+		booleanFilter.addRequiredTerm(Field.HIDDEN, false);
+
+		_filterByBasePrice(booleanFilter, searchContext);
+		_filterByDefinitionLinks(booleanFilter, searchContext);
+		_filterByExcludedCPDefinitionId(booleanFilter, searchContext);
+		_filterByPublished(booleanFilter, searchContext);
+		_filterByStatuses(booleanFilter, searchContext);
+		_filterBySubscriptionEnabled(booleanFilter, searchContext);
+
+		if (GetterUtil.getBoolean(searchContext.getAttribute("secure"))) {
+			_filterByAccountGroupIds(booleanFilter, searchContext);
+			_filterByCommerceChannelId(booleanFilter, searchContext);
+		}
+
 		if (_isIndexersSuppressed(searchContext)) {
-			_addCommerceCatalogIdFilterClauses(booleanFilter, searchContext);
+			_filterByCommerceCatalogIds(booleanFilter, searchContext);
 		}
 	}
 
-	private void _addCommerceCatalogIdFilterClauses(
+	private void _filterByAccountGroupIds(
 		BooleanFilter booleanFilter, SearchContext searchContext) {
 
-		TermsFilter termsFilter = new TermsFilter("commerceCatalogId");
+		BooleanFilter accountGroupsBooleanFilter = new BooleanFilter();
+
+		BooleanFilter accountGroupsFilterEnableBooleanFilter =
+			new BooleanFilter();
+
+		accountGroupsFilterEnableBooleanFilter.addTerm(
+			CPField.ACCOUNT_GROUP_FILTER_ENABLED, Boolean.TRUE.toString(),
+			BooleanClauseOccur.MUST);
+
+		long[] accountGroupIds = GetterUtil.getLongValues(
+			searchContext.getAttribute("commerceAccountGroupIds"), null);
+
+		if ((accountGroupIds != null) && (accountGroupIds.length > 0)) {
+			BooleanFilter accountGroupIdsBooleanFilter = new BooleanFilter();
+
+			for (long accountGroupId : accountGroupIds) {
+				Filter termFilter = new TermFilter(
+					"commerceAccountGroupIds", String.valueOf(accountGroupId));
+
+				accountGroupIdsBooleanFilter.add(
+					termFilter, BooleanClauseOccur.SHOULD);
+			}
+
+			accountGroupsFilterEnableBooleanFilter.add(
+				accountGroupIdsBooleanFilter, BooleanClauseOccur.MUST);
+		}
+		else {
+			accountGroupsFilterEnableBooleanFilter.addTerm(
+				"commerceAccountGroupIds", "-1", BooleanClauseOccur.MUST);
+		}
+
+		accountGroupsBooleanFilter.add(
+			accountGroupsFilterEnableBooleanFilter, BooleanClauseOccur.SHOULD);
+		accountGroupsBooleanFilter.addTerm(
+			CPField.ACCOUNT_GROUP_FILTER_ENABLED, Boolean.FALSE.toString(),
+			BooleanClauseOccur.SHOULD);
+
+		boolean ignoreAccountGroup = GetterUtil.getBoolean(
+			searchContext.getAttribute("ignoreCommerceAccountGroup"));
+
+		if (!ignoreAccountGroup) {
+			booleanFilter.add(
+				accountGroupsBooleanFilter, BooleanClauseOccur.MUST);
+		}
+	}
+
+	private void _filterByBasePrice(
+		BooleanFilter booleanFilter, SearchContext searchContext) {
+
+		String[] basePriceRanges = GetterUtil.getStringValues(
+			searchContext.getAttribute(CPField.BASE_PRICE));
+
+		if (basePriceRanges.length < 1) {
+			return;
+		}
+
+		BooleanFilter priceRangeBooleanFilter = new BooleanFilter();
+
+		for (String basePriceRange : basePriceRanges) {
+			String[] basePriceRangeParts = RangeParserUtil.parserRange(
+				basePriceRange);
+
+			RangeTermFilter rangeTermFilter = new RangeTermFilter(
+				CPField.BASE_PRICE, true, true, basePriceRangeParts[0],
+				basePriceRangeParts[1]);
+
+			priceRangeBooleanFilter.add(
+				rangeTermFilter, BooleanClauseOccur.SHOULD);
+		}
+
+		booleanFilter.add(priceRangeBooleanFilter, BooleanClauseOccur.MUST);
+	}
+
+	private void _filterByCommerceCatalogIds(
+		BooleanFilter booleanFilter, SearchContext searchContext) {
 
 		long[] commerceCatalogIds = _getCommerceCatalogIds(searchContext);
 
-		if (commerceCatalogIds.length == 0) {
-			termsFilter.addValue("-1");
+		if (commerceCatalogIds.length > 0) {
+			TermsFilter termsFilter = new TermsFilter("commerceCatalogId");
+
+			termsFilter.addValues(ArrayUtil.toStringArray(commerceCatalogIds));
+
+			booleanFilter.add(termsFilter, BooleanClauseOccur.MUST);
 		}
 		else {
-			termsFilter.addValues(ArrayUtil.toStringArray(commerceCatalogIds));
+			long[] groupIds = searchContext.getGroupIds();
+
+			if ((groupIds == null) || (groupIds.length == 0)) {
+				booleanFilter.addTerm(
+					Field.GROUP_ID, "-1", BooleanClauseOccur.MUST);
+			}
+		}
+	}
+
+	private void _filterByCommerceChannelId(
+		BooleanFilter booleanFilter, SearchContext searchContext) {
+
+		BooleanFilter commerceChannelBooleanFilter = new BooleanFilter();
+
+		BooleanFilter commerceChannelFilterEnableBooleanFilter =
+			new BooleanFilter();
+
+		commerceChannelFilterEnableBooleanFilter.addTerm(
+			CPField.CHANNEL_FILTER_ENABLED, Boolean.TRUE.toString(),
+			BooleanClauseOccur.MUST);
+
+		long commerceChannelId = GetterUtil.getLong(
+			searchContext.getAttribute("commerceChannelGroupId"));
+
+		if (commerceChannelId > 0) {
+			commerceChannelFilterEnableBooleanFilter.addTerm(
+				CPField.COMMERCE_CHANNEL_GROUP_IDS,
+				String.valueOf(commerceChannelId), BooleanClauseOccur.MUST);
+		}
+		else {
+			commerceChannelFilterEnableBooleanFilter.addTerm(
+				CPField.COMMERCE_CHANNEL_GROUP_IDS, "-1",
+				BooleanClauseOccur.MUST);
 		}
 
-		booleanFilter.add(termsFilter, BooleanClauseOccur.MUST);
+		commerceChannelBooleanFilter.add(
+			commerceChannelFilterEnableBooleanFilter,
+			BooleanClauseOccur.SHOULD);
+		commerceChannelBooleanFilter.addTerm(
+			CPField.CHANNEL_FILTER_ENABLED, Boolean.FALSE.toString(),
+			BooleanClauseOccur.SHOULD);
+
+		booleanFilter.add(
+			commerceChannelBooleanFilter, BooleanClauseOccur.MUST);
+	}
+
+	private void _filterByDefinitionLinks(
+		BooleanFilter booleanFilter, SearchContext searchContext) {
+
+		String definitionLinkType = GetterUtil.getString(
+			searchContext.getAttribute("definitionLinkType"));
+
+		long definitionLinkCPDefinitionId = GetterUtil.getLong(
+			searchContext.getAttribute("definitionLinkCPDefinitionId"));
+
+		if (Validator.isNotNull(definitionLinkType) &&
+			(definitionLinkCPDefinitionId > 0)) {
+
+			TermsFilter termsFilter = new TermsFilter(definitionLinkType);
+
+			termsFilter.addValue(String.valueOf(definitionLinkCPDefinitionId));
+
+			booleanFilter.add(termsFilter, BooleanClauseOccur.MUST);
+		}
+	}
+
+	private void _filterByExcludedCPDefinitionId(
+		BooleanFilter booleanFilter, SearchContext searchContext) {
+
+		String excludedCPDefinitionId = GetterUtil.getString(
+			searchContext.getAttribute("excludedCPDefinitionId"));
+
+		if (Validator.isNotNull(excludedCPDefinitionId)) {
+			booleanFilter.addTerm(
+				Field.ENTRY_CLASS_PK, excludedCPDefinitionId,
+				BooleanClauseOccur.MUST_NOT);
+		}
+	}
+
+	private void _filterByPublished(
+		BooleanFilter booleanFilter, SearchContext searchContext) {
+
+		boolean published = GetterUtil.getBoolean(
+			searchContext.getAttribute(CPField.PUBLISHED));
+
+		if (published) {
+			booleanFilter.addRequiredTerm(CPField.PUBLISHED, true);
+		}
+	}
+
+	private void _filterByStatuses(
+		BooleanFilter booleanFilter, SearchContext searchContext) {
+
+		int[] statuses = GetterUtil.getIntegerValues(
+			searchContext.getAttribute(Field.STATUS));
+
+		if (ArrayUtil.isEmpty(statuses)) {
+			int status = GetterUtil.getInteger(
+				searchContext.getAttribute(Field.STATUS),
+				WorkflowConstants.STATUS_APPROVED);
+
+			statuses = new int[] {status};
+		}
+
+		if (!ArrayUtil.contains(statuses, WorkflowConstants.STATUS_ANY)) {
+			TermsFilter termsFilter = new TermsFilter(Field.STATUS);
+
+			termsFilter.addValues(ArrayUtil.toStringArray(statuses));
+
+			booleanFilter.add(termsFilter, BooleanClauseOccur.MUST);
+		}
+	}
+
+	private void _filterBySubscriptionEnabled(
+		BooleanFilter booleanFilter, SearchContext searchContext) {
+
+		boolean subscriptionEnabled = GetterUtil.getBoolean(
+			searchContext.getAttribute(CPField.SUBSCRIPTION_ENABLED));
+
+		if (subscriptionEnabled) {
+			booleanFilter.addRequiredTerm(CPField.SUBSCRIPTION_ENABLED, true);
+		}
 	}
 
 	private long[] _getCommerceCatalogIds(SearchContext searchContext) {
