+++ modules/apps/commerce/commerce-product-service/src/main/java/com/liferay/commerce/product/internal/search/spi/model/result/contributor/CPDefinitionModelSummaryContributor.java
@@ -1,0 +1,45 @@
+/**
+ * SPDX-FileCopyrightText: (c) 2023 Liferay, Inc. https://liferay.com
+ * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
+ */
+
+package com.liferay.commerce.product.internal.search.spi.model.result.contributor;
+
+import com.liferay.petra.string.StringPool;
+import com.liferay.portal.kernel.search.Document;
+import com.liferay.portal.kernel.search.Field;
+import com.liferay.portal.kernel.search.Summary;
+import com.liferay.portal.search.spi.model.result.contributor.ModelSummaryContributor;
+
+import java.util.Locale;
+
+import org.osgi.service.component.annotations.Component;
+
+/**
+ * @author Brian I. Kim
+ */
+@Component(
+	property = "indexer.class.name=com.liferay.commerce.product.model.CPDefinition",
+	service = ModelSummaryContributor.class
+)
+public class CPDefinitionModelSummaryContributor
+	implements ModelSummaryContributor {
+
+	@Override
+	public Summary getSummary(
+		Document document, Locale locale, String snippet) {
+
+		String prefix = Field.SNIPPET + StringPool.UNDERLINE;
+
+		String title = document.get(prefix + Field.NAME, Field.NAME);
+		String content = document.get(
+			prefix + Field.DESCRIPTION, Field.DESCRIPTION);
+
+		Summary summary = new Summary(title, content);
+
+		summary.setMaxContentLength(200);
+
+		return summary;
+	}
+
+}
